requirements.txt
requests
setuptools
notion-client
python-telegram-bot